requirements.txt
requests>=2.31.0
PySide6>=6.7.0
Pillow>=10.0.0
python-json-logger>=2.0.7
